Placate clippy

+++ musq/src/sqlite/statement/virtual.rs
@@ -168,7 +168,7 @@ fn prepare(
                 conn,
                 query_ptr,
                 query_len,
-                flags as u32,
+                flags,
                 &mut statement_handle,
                 &mut tail,
             )
